+++ .pre-commit-config.yaml
@@ -5,3 +5,9 @@ repos:
     -   id: check-yaml
     -   id: end-of-file-fixer
     -   id: trailing-whitespace
+-   repo: https://github.com/pre-commit/mirrors-prettier
+    rev: v2.3.2
+    hooks:
+    -   id: prettier
+        additional_dependencies: ['prettier@2.3.2']
+        files: \.(ts|tsx|js|jsx|json|css|scss|md|yaml|yml)$
